Reverting MOAITimer ( randomly reverted in last commit ) Fixing soak on windows Fixing compile warning in prop

+++ src/moaicore/MOAIProp.cpp
@@ -563,7 +563,7 @@ void MOAIProp::Draw ( int subPrimID, bool reload ) {
 	
 	if ( this->mGrid ) {
 	
-		if ( subPrimID == MOAIProp::NO_SUBPRIM_ID ) {
+		if ( (u32)subPrimID == MOAIProp::NO_SUBPRIM_ID ) {
 	
 			MOAICellCoord c0;
 			MOAICellCoord c1;
